pyproject.toml
[project]
name = "readme-credly-badges"
version = "0.1.0"
description = "A Python GitHub Action that auto-updates the README with the Credly badge details."
readme = "README.md"
keywords = [  ]

license = "MIT"
authors = [ { name = "jd-35656", email = "jitesh.sahani@outlook.com" } ]

requires-python = ">=3.8"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [ "pygithub>=2.6.1", "requests>=2.32.3" ]

urls.Documentation = "https://github.com/jd-35656/readme-credly-badges#readme"
urls.Issues = "https://github.com/jd-35656/readme-credly-badges/issues"
urls.Source = "https://github.com/jd-35656/readme-credly-badges"

[tool.ruff]
target-version = "py311"
line-length = 121
src = [ "src" ]

# Formatting options
format.indent-style = "space"
format.quote-style = "double"
format.skip-magic-trailing-comma = false

# Linting options
lint.select = [
  "A",
  "ARG",
  "B",
  "C4",
  "C9",
  "DTZ",
  "E",
  "F",
  "I",
  "ISC",
  "N",
  "PERF",
  "PGH",
  "PIE",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "PYI",
  "Q",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "T20",
  "UP",
  "W",
]

lint.per-file-ignores."src/readme_credly_badges/adapter/github_repo.py" = [
  "PLR0913",
]
lint.per-file-ignores."tests/*" = [ "N806", "S101", "S106" ]

lint.isort.known-first-party = [ "readme_credly_badges" ]
lint.mccabe.max-complexity = 15

[tool.codespell]
skip = ".git,venv,__pycache__,.mypy_cache,.ruff_cache"
check-hidden = true
quiet-level = 2

[tool.pytest.ini_options]
minversion = "8.0"
log_cli = true
log_cli_level = "INFO"
testpaths = [ "tests" ]
pythonpath = [ "src" ]
addopts = [ "-ra", "--strict-config", "--strict-markers" ]

[tool.coverage.run]
source_pkgs = [ "readme_credly_badges", "tests" ]
branch = true
parallel = true

[tool.coverage.paths]
readme_credly_badges = [
  "src/readme_credly_badges",
  "*/readme-credly-badges/src/readme_credly_badges",
]
tests = [ "tests", "*/readme-credly-badges/tests" ]

[tool.coverage.report]
exclude_lines = [ "no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:" ]

[tool.towncrier]
directory = "changelog.d"
filename = "docs/changelog.md"
package = "readme_credly_badges"
start_string = "<!-- towncrier release notes start -->\n"
title_format = "## [{version}](https://github.com/jd-35656/readme-credly-badges/tree/{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/jd-35656/readme-credly-badges/issues/{issue})"
underlines = [ "", "", "" ]

[tool.nox.dependency-group]
test = [ "pytest >=8.3.5", "pytest-cov >=6.1.1", "toml" ]
docs = [ "towncrier" ]
